refactor: complete modularization of src components and store
- Updated TASK004 status to completed in documentation.
- Refactored `BricksInstanced` into a modular structure with hooks and utilities.
- Created dedicated components for UI panels: `ScorePanel`, `StatsPanel`, `UpgradesPanel`, `AchievementsPanel`, and `Controls`.
- Removed old UI CSS and integrated styles into new UI components.
- Implemented keyboard shortcuts for game controls in a separate hook.
- Added unit tests for the `getDamageColor` utility function.
