+++ memory/tasks/_index.md
@@ -7,12 +7,12 @@
 ## IN PROGRESS
 
 - [TASK002] Performance Profiling & UX Polish - Profile rendering and optimize GameScene/Ball updates; polish UI
-- [TASK004] Refactor: Modularize `src/` (Store & Components) - Modularize store, BricksInstanced, and UI without behavior change
 
 ## COMPLETED
 
 - [TASK001] Wave & Meta Progression Implementation - Implemented waves, achievements, partial persistence (meta-only), and UI updates (see tests and store updates)
 - [TASK003] Investigate Persist/Rehydrate Race - Investigated and implemented companion metadata fix; tests updated and passing
+- [TASK004] Refactor: Modularize `src/` (Store & Components) - Modularized store, BricksInstanced, and UI without behavior change
 
 ## ABANDONED
