+++ src/components/bricks/useInstancedBricks.ts
@@ -1,25 +1,14 @@
 import { useCallback, useLayoutEffect, useRef } from 'react';
-import { Object3D, Color, type InstancedMesh } from 'three';
+import { Color, Object3D, type InstancedMesh } from 'three';
 import type { ThreeEvent } from '@react-three/fiber';
-import type { Brick } from '../store/gameStore';
-import { BRICK_SIZE } from '../engine/collision';
-import { getBrickFromInstance } from '../engine/picking';
+import { getBrickFromInstance } from '../../engine/picking';
+import type { Brick } from '../../store/gameStore';
+import { getDamageColor } from './utils';
 
 const tempObject = new Object3D();
 const tempColor = new Color();
 
-const getDamageColor = (brick: Brick, isHovered: boolean) => {
-  if (isHovered) return '#FFFFFF';
-  const healthRatio = brick.health / brick.maxHealth;
-  if (healthRatio > 0.5) return brick.color;
-  return `hsl(${Math.floor(healthRatio * 60)}, 80%, 50%)`;
-};
-
-interface BricksInstancedProps {
-  bricks: Brick[];
-}
-
-export function BricksInstanced({ bricks }: BricksInstancedProps) {
+export const useInstancedBricks = (bricks: Brick[]) => {
   const meshRef = useRef<InstancedMesh>(null);
   const hoveredIndexRef = useRef<number | null>(null);
   const hoveredBrickIdRef = useRef<string | null>(null);
@@ -129,22 +118,9 @@ export function BricksInstanced({ bricks }: BricksInstancedProps) {
     clearHoveredInstance();
   }, [clearHoveredInstance]);
 
-  return (
-    <instancedMesh
-      ref={meshRef}
-      args={[undefined, undefined, bricks.length]}
-      castShadow
-      receiveShadow
-      onPointerMove={handlePointerMove}
-      onPointerOut={handlePointerOut}
-    >
-      <boxGeometry args={[BRICK_SIZE.x, BRICK_SIZE.y, BRICK_SIZE.z]} />
-      <meshStandardMaterial
-        emissiveIntensity={0.1}
-        metalness={0.3}
-        roughness={0.7}
-        transparent
-      />
-    </instancedMesh>
-  );
-}
+  return {
+    meshRef,
+    handlePointerMove,
+    handlePointerOut,
+  };
+};
