+++ src/components/bricks/utils.ts
@@ -1,0 +1,8 @@
+import type { Brick } from '../../store/gameStore';
+
+export const getDamageColor = (brick: Brick, isHovered: boolean) => {
+  if (isHovered) return '#FFFFFF';
+  const healthRatio = brick.health / brick.maxHealth;
+  if (healthRatio > 0.5) return brick.color;
+  return `hsl(${Math.floor(healthRatio * 60)}, 80%, 50%)`;
+};
